1. Modified Libraries in pedal_pico  a. Removed Process for Normalizing Biasing for Audio Input  b. Removed Correction of Audio Input on Each Library
2. Modified util_pedal_pico
 a. Added Process for Normalizing and Biasing for Audio Input
 b. Added Correction of Audio Input on Each Library

3. Modified pedal_multi to Fit with Modification of Libraries in pedal_pico and util_pedal_pico

+++ include/util_pedal_pico.h
@@ -152,7 +152,7 @@ extern int32 util_pedal_pico_ex_table_log_2[];
 extern int32 util_pedal_pico_ex_table_power_1[];
 
 /* Functions */
-void (*util_pedal_pico_process)(uint16, uint16, uint16, uchar8); // Pointer Needed to Be Assigned
+void (*util_pedal_pico_process)(int32, uint16, uint16, uchar8); // Pointer Needed to Be Assigned
 void util_pedal_pico_set_sys_clock_115200khz();
 void util_pedal_pico_set_pwm_28125hz(pwm_config* ptr_config);
 util_pedal_pico* util_pedal_pico_init(uchar8 gpio_1, uchar8 gpio_2);
